breweries {mapview2}	R Documentation

Selected breweries in Franconia

Description

Selected breweries in Franconia

Format

sf feature collection POINT

Details

This dataset contains selected breweries in Franconia. It is partly a subset of a larger database that was compiled by students at the University of Marburg for a seminar called "The Geography of Beer: sustainability in the food industry" and partly consists of breweries downloaded from https://www.bierwandern.de/inhalt/brauereiliste.html with the kind permission of Rainer Kastl. Note that use of these data is restricted to non-commercial use and that they are explixitly excluded from the GPL lincense that mapview is licensed under.

[Package mapview2 version 2.10.4.9005 Index]